# openapi.yaml
openapi: 3.0.3
info:
  title: API REST para Sistema de Noticias
  description: API para gestión de noticias, comentarios, usuarios y categorías
  version: 1.0.0
servers:
  - url: http://api.example.com/v1
    description: Servidor de producción
  - url: http://localhost:8080/v1
    description: Servidor de desarrollo
paths:
...
  /usuarios:
    post:
      summary: Crear un usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioInput'
      responses:
        '201':
          description: Usuario creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        '400':
          description: Solicitud incorrecta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /usuarios/{usuario-id}:
    get:
      summary: Publicaciones de un usuario
      parameters:
        - $ref: '#/components/parameters/UsuarioId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Noticia'
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Actualizar datos del usuario
      parameters:
        - $ref: '#/components/parameters/UsuarioId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioInput'
      responses:
        '200':
          description: Usuario actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        '400':
          description: Solicitud incorrecta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Borrar un usuario
      parameters:
        - $ref: '#/components/parameters/UsuarioId'
      responses:
        '204':
          description: Usuario eliminado (sin contenido)
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Noticia:
      type: object
      properties:
        id:
          type: integer
          format: int64
        titulo:
          type: string
        contenido:
          type: string
        fecha:
          type: string
          format: date-time
        autor:
          $ref: '#/components/schemas/Usuario'
        categoria:
          $ref: '#/components/schemas/Categoria'
        comentarios:
          type: array
          items:
            $ref: '#/components/schemas/Comentario'
      required:
        - id
        - titulo
        - contenido
        - fecha
        - autor
...
    Comentario:
      type: object
      properties:
        id:
          type: integer
          format: int64
        contenido:
          type: string
        fecha:
          type: string
          format: date-time
        autor:
          $ref: '#/components/schemas/Usuario'
        noticiaId:
          type: integer
          format: int64
      required:
        - id
        - contenido
        - fecha
        - autor
        - noticiaId
...
    Usuario:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        email:
          type: string
          format: email
        fechaRegistro:
          type: string
          format: date-time
      required:
        - id
        - nombre
        - email
        - fechaRegistro
    UsuarioInput:
      type: object
      properties:
        nombre:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
      required:
        - nombre
        - email
        - password
    Categoria:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        descripcion:
          type: string
      required:
        - id
        - nombre
...
    Error:
      type: object
      properties:
        codigo:
          type: integer
        mensaje:
          type: string
        detalles:
          type: string
      required:
        - codigo
        - mensaje
  parameters:
...
    UsuarioId:
      name: usuario-id
      in: path
      description: ID del usuario
      required: true
      schema:
        type: integer
        format: int64
...